requirements.txt
hvac
boto3
requests
sendgrid
python-dotenv
pymongo
pikepdf
mail-parser
pyotp